# backend/openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/admin/ledger/balances": {
      "get": {
        "operationId": "LedgerAdminController_balances",
        "parameters": [
          {
            "name": "page",
            "required": false,
            "in": "query",
            "schema": {
              "minimum": 1,
              "default": 1,
              "type": "number"
            }
          },
          {
            "name": "pageSize",
            "required": false,
            "in": "query",
            "schema": {
              "minimum": 1,
              "maximum": 100,
              "default": 20,
              "type": "number"
            }
          },
          {
            "name": "search",
            "required": false,
            "in": "query",
            "description": "Recherche sur le code membre, le nom ou le prénom.",
            "schema": {
              "maxLength": 120,
              "type": "string"
            }
          },
          {
            "name": "status",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string",
              "enum": [
                "REGISTERED",
                "ACTIVE",
                "INACTIVE"
              ]
            }
          },
          {
            "name": "withBalanceOnly",
            "required": false,
            "in": "query",
            "description": "N’afficher que les membres dont le solde est strictement positif.",
            "schema": {
              "type": "boolean"
            }
          },
          {
            "name": "sort",
            "required": false,
            "in": "query",
            "schema": {
              "default": "balanceDt",
              "type": "string",
              "enum": [
                "memberCode",
                "lastName",
                "balanceDt"
              ]
            }
          },
          {
            "name": "direction",
            "required": false,
            "in": "query",
            "schema": {
              "default": "desc",
              "type": "string",
              "enum": [
                "asc",
                "desc"
              ]
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BalancePageDto"
                }
              }
            }
          }
        },
        "summary": "Registre des soldes par membre (recherche, filtre d’état, tri, pagination) + somme des soldes filtrés.",
        "tags": [
          "ledger-admin"
        ]
      }
    },
    "/admin/ledger/movements": {
      "get": {
        "operationId": "LedgerAdminController_movements",
        "parameters": [
          {
            "name": "page",
            "required": false,
            "in": "query",
            "schema": {
              "minimum": 1,
              "default": 1,
              "type": "number"
            }
          },
          {
            "name": "pageSize",
            "required": false,
            "in": "query",
            "schema": {
              "minimum": 1,
              "maximum": 100,
              "default": 20,
              "type": "number"
            }
          },
          {
            "name": "search",
            "required": false,
            "in": "query",
            "description": "Recherche sur le code membre, le nom ou le prénom.",
            "schema": {
              "maxLength": 120,
              "type": "string"
            }
          },
          {
            "name": "memberId",
            "required": false,
            "in": "query",
            "description": "Restreindre à un membre.",
            "schema": {
              "minimum": 1,
              "type": "number"
            }
          },
          {
            "name": "type",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string",
              "enum": [
                "ECARD_CREATION",
                "ECARD_REFUND",
                "COMMISSION",
                "ACTIVATION",
                "ADMIN_ADJUSTMENT",
                "ADMIN_GENESIS"
              ]
            }
          },
          {
            "name": "from",
            "required": false,
            "in": "query",
            "description": "Mouvements à partir de cette date (incluse), au format ISO.",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "to",
            "required": false,
            "in": "query",
            "description": "Mouvements jusqu’à cette date (incluse), au format ISO.",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MovementPageDto"
                }
              }
            }
          }
        },
        "summary": "Journal GLOBAL des mouvements de solde (type, montant signé, solde après, source, date).",
        "tags": [
          "ledger-admin"
        ]
      }
    },
    "/admin/ledger/members/{memberId}/adjustment": {
      "post": {
        "operationId": "LedgerAdminController_adjust",
        "parameters": [
          {
            "name": "memberId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AdjustBalanceDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LedgerEntryResponseDto"
                }
              }
            }
          }
        },
        "summary": "Ajustement manuel du solde en DT (motif obligatoire, tracé)",
        "tags": [
          "ledger-admin"
        ]
      }
    },
    "/admin/ledger/members/{memberId}/genesis": {
      "post": {
        "operationId": "LedgerAdminController_genesis",
        "parameters": [
          {
            "name": "memberId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/GenesisBalanceDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LedgerEntryResponseDto"
                }
              }
            }
          }
        },
        "summary": "Génération de solde en DT ex nihilo (amorçage / promo, ADMIN_GENESIS) — motif OBLIGATOIRE, tracé.",
        "tags": [
          "ledger-admin"
        ]
      }
    },
    "/admin/ledger/members/{memberId}/balance": {
      "get": {
        "operationId": "LedgerAdminController_balance",
        "parameters": [
          {
            "name": "memberId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MemberBalanceResponseDto"
                }
              }
            }
          }
        },
        "summary": "Solde courant d’un membre (DT)",
        "tags": [
          "ledger-admin"
        ]
      }
    },
    "/admin/ledger/members/{memberId}/history": {
      "get": {
        "operationId": "LedgerAdminController_history",
        "parameters": [
          {
            "name": "memberId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          },
          {
            "name": "page",
            "required": false,
            "in": "query",
            "schema": {
              "minimum": 1,
              "default": 1,
              "type": "number"
            }
          },
          {
            "name": "pageSize",
            "required": false,
            "in": "query",
            "schema": {
              "minimum": 1,
              "maximum": 100,
              "default": 20,
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LedgerHistoryPageDto"
                }
              }
            }
          }
        },
        "summary": "Historique paginé des mouvements de solde d’un membre (DT)",
        "tags": [
          "ledger-admin"
        ]
      }
    },
...
  },
  "info": {
    "title": "Najah Partners API",
    "description": "API backend — plateforme MLM (BV, e-cards, commissions)",
    "version": "1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      }
    },
    "schemas": {
...
      "BalanceRowDto": {
        "type": "object",
        "properties": {
          "memberId": {
            "type": "number"
          },
          "memberCode": {
            "type": "string"
          },
          "firstName": {
            "type": "string"
          },
          "lastName": {
            "type": "string"
          },
          "status": {
            "enum": [
              "REGISTERED",
              "ACTIVE",
              "INACTIVE"
            ],
            "type": "string"
          },
          "balanceDt": {
            "type": "string",
            "example": "1250.500",
            "description": "DINARS — solde courant."
          },
          "movementCount": {
            "type": "number",
            "description": "Mouvements enregistrés pour ce membre."
          },
          "lastMovementAt": {
            "format": "date-time",
            "type": "string",
            "nullable": true,
            "description": "Date du dernier mouvement. `null` = aucun mouvement : un solde à 0 sans histoire, ce qui est le cas normal d’un INSCRIT (payer par e-card n’écrit RIEN au grand livre — D-025)."
          }
        },
        "required": [
          "memberId",
          "memberCode",
          "firstName",
          "lastName",
          "status",
          "balanceDt",
          "movementCount"
        ]
      },
      "BalancePageDto": {
        "type": "object",
        "properties": {
          "items": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/BalanceRowDto"
            }
          },
          "total": {
            "type": "number"
          },
          "page": {
            "type": "number"
          },
          "pageSize": {
            "type": "number"
          },
          "totalBalanceDt": {
            "type": "string",
            "example": "31200.500",
            "description": "DINARS — somme des soldes de TOUS les membres correspondant au filtre (pas seulement de la page)."
          }
        },
        "required": [
          "items",
          "total",
          "page",
          "pageSize",
          "totalBalanceDt"
        ]
      },
      "MovementMemberRefDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "number"
          },
          "memberCode": {
            "type": "string"
          },
          "firstName": {
            "type": "string"
          },
          "lastName": {
            "type": "string"
          }
        },
        "required": [
          "id",
          "memberCode",
          "firstName",
          "lastName"
        ]
      },
      "MovementRowDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "number"
          },
          "member": {
            "$ref": "#/components/schemas/MovementMemberRefDto"
          },
          "type": {
            "enum": [
              "ECARD_CREATION",
              "ECARD_REFUND",
              "COMMISSION",
              "ACTIVATION",
              "ADMIN_ADJUSTMENT",
              "ADMIN_GENESIS"
            ],
            "type": "string",
            "description": "Il n’existe PAS de `ECARD_USE` (D-025) : consommer une e-card ne bouge aucun solde, donc n’écrit rien ici."
          },
          "amountDt": {
            "type": "string",
            "example": "-100.000",
            "description": "DINARS, SIGNÉ : + crédit, − débit."
          },
          "balanceAfterDt": {
            "type": "string",
            "example": "1150.500",
            "description": "DINARS — solde APRÈS le mouvement."
          },
          "ecardId": {
            "type": "number",
            "nullable": true,
            "description": "E-card à l’origine du mouvement (id, jamais le code)."
          },
          "commissionId": {
            "type": "number",
            "nullable": true
          },
          "reason": {
            "type": "string",
            "nullable": true,
            "description": "Motif — OBLIGATOIRE pour un ajustement comme pour une genèse."
          },
          "createdAt": {
            "format": "date-time",
            "type": "string"
          }
        },
        "required": [
          "id",
          "member",
          "type",
          "amountDt",
          "balanceAfterDt",
          "createdAt"
        ]
      },
      "MovementPageDto": {
        "type": "object",
        "properties": {
          "items": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/MovementRowDto"
            }
          },
          "total": {
            "type": "number"
          },
          "page": {
            "type": "number"
          },
          "pageSize": {
            "type": "number"
          },
          "netAmountDt": {
            "type": "string",
            "example": "4300.000",
            "description": "DINARS — somme SIGNÉE des mouvements filtrés (pas seulement de la page)."
          }
        },
        "required": [
          "items",
          "total",
          "page",
          "pageSize",
          "netAmountDt"
        ]
      },
      "AdjustBalanceDto": {
        "type": "object",
        "properties": {
          "amountDt": {
            "type": "number",
            "description": "Montant signé en DINARS (+ crédit / − débit), non nul. 3 décimales au maximum (le millime).",
            "example": -500
          },
          "reason": {
            "type": "string",
            "maxLength": 500,
            "description": "Motif obligatoire (tracé dans le grand livre et l’audit).",
            "example": "Régularisation suite à une erreur de saisie"
          }
        },
        "required": [
          "amountDt",
          "reason"
        ]
      },
      "LedgerEntryResponseDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "number"
          },
          "memberId": {
            "type": "number"
          },
          "type": {
            "enum": [
              "ECARD_CREATION",
              "ECARD_REFUND",
              "COMMISSION",
              "ACTIVATION",
              "ADMIN_ADJUSTMENT",
              "ADMIN_GENESIS"
            ],
            "type": "string",
            "description": "Il n’existe PAS de valeur `ECARD_USE` (D-025) : consommer une e-card ne bouge aucun solde, donc n’écrit aucune ligne ici."
          },
          "amountDt": {
            "type": "string",
            "example": "-100.000",
            "description": "DINARS, SIGNÉ : + crédit, − débit."
          },
          "balanceAfterDt": {
            "type": "string",
            "example": "1150.500",
            "description": "DINARS — solde du membre APRÈS le mouvement."
          },
          "ecardId": {
            "type": "number",
            "nullable": true,
            "description": "E-card à l’origine du mouvement."
          },
          "commissionId": {
            "type": "number",
            "nullable": true,
            "description": "Run de commissions à l’origine du mouvement."
          },
          "reason": {
            "type": "string",
            "nullable": true,
            "description": "Motif — OBLIGATOIRE pour un ADMIN_ADJUSTMENT."
          },
          "createdAt": {
            "format": "date-time",
            "type": "string"
          }
        },
        "required": [
          "id",
          "memberId",
          "type",
          "amountDt",
          "balanceAfterDt",
          "createdAt"
        ]
      },
      "GenesisBalanceDto": {
        "type": "object",
        "properties": {
          "amountDt": {
            "type": "number",
            "minimum": 0.001,
            "description": "Montant en DINARS à générer (amorçage réseau / promo), strictement positif. 3 décimales au maximum.",
            "example": 2200
          },
          "reason": {
            "type": "string",
            "maxLength": 500,
            "description": "Motif OBLIGATOIRE (tracé dans le grand livre et l’audit).",
            "example": "Amorçage réseau — dotation initiale"
          }
        },
        "required": [
          "amountDt",
          "reason"
        ]
      },
      "MemberBalanceResponseDto": {
        "type": "object",
        "properties": {
          "memberId": {
            "type": "number"
          },
          "balanceDt": {
            "type": "string",
            "example": "1250.500",
            "description": "DINARS — solde courant."
          }
        },
        "required": [
          "memberId",
          "balanceDt"
        ]
      },
      "LedgerHistoryPageDto": {
        "type": "object",
        "properties": {
          "items": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/LedgerEntryResponseDto"
            }
          },
          "total": {
            "type": "number"
          },
          "page": {
            "type": "number"
          },
          "pageSize": {
            "type": "number"
          }
        },
        "required": [
          "items",
          "total",
          "page",
          "pageSize"
        ]
      },
...
    }
  }
}